requirements.txt
appdirs==1.4.0
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.8
packaging==16.8
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
